requirements.txt
Django
PyYAML
South
anyjson
billiard
dj-database-url
dj-static
django-cors-headers
django-taggit
django-toolbelt
django-voting
djangorestframework
gunicorn
kombu
pystache
pytz
static
nltk
django-extensions